fix issue #441
When trying to replicate issue #441, we found out that this issue did
not appear in an older version of the code. Cdn-admin had the code
from commit 5fdb4d7 from 8 Augustus 2024. There the issue did not
occur and it works as expected.
However, at current dev at commit ec0c202, this issue occurs like
Emil described. I looked at what the difference was between these
commits related to line 994 in yolo_utils.py, which is where the error
message points to. This line was changed in commit 3ce6150 (Fix missing
species_id issue) and it is unclear to me why this change was made in
the code.
But it is also unclear to me why other parts of this same function
(frame_aggregation) did already use name[0] instead of name before
this commit. So it seemed like this was the only part where we use
name instead of name[0]. So maybe that was the reason why it was
changed?

However, if I change it back and partly revert that commit, so that we
just use name instead of name[0], the bug disappears and the code runs.
Since that solves the problem, that is what this commit does.
However, this has the risk of introducing new bugs somewhere else,
since it is unclear to me why this code was changed in the first place.