[project.py] set yolo settings dependent on registry
We want to track the training process either with WandB or with MLflow.
Yolo from Ultralytics has build in support for both. But to use the
tracking with that platform, you need to set the settings for that
platform to True.
See Ultralytics documentation:
https://github.com/ultralytics/ultralytics/blob/21162bd870444550286983a601afbfb142f4c198/docs/en/integrations/mlflow.md

I don't understand how the code could have worked together with MLflow
before, since I think that this was missing.
However, that we always set this setting to WandB explains why we always
got WandB information in the printed output of the training, even when
we tried to use MLflow.

I think that this change should fix issue #444.